+++ android/app/build.gradle
@@ -77,7 +77,9 @@ android {
     ndkVersion rootProject.ext.ndkVersion
     buildToolsVersion rootProject.ext.buildToolsVersion
     compileSdk rootProject.ext.compileSdkVersion
-
+    lintOptions {
+        disable 'RestrictedApi'  // 이 오류를 무시하도록 설정
+    }
     namespace "com.gnu.mojadol"
     defaultConfig {
         applicationId "com.gnu.mojadol"
@@ -105,6 +107,11 @@ android {
             minifyEnabled enableProguardInReleaseBuilds
             proguardFiles getDefaultProguardFile("proguard-android.txt"), "proguard-rules.pro"
         }
+    }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_17  // Java 17 사용
+        targetCompatibility JavaVersion.VERSION_17  // Java 17 사용
     }
 }
