+++ android/app/src/main/AndroidManifest.xml
@@ -5,9 +5,6 @@
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
     <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
 
-
-
-
     <application
       android:name=".MainApplication"
       android:label="@string/app_name"
