+++ src/screens/LoginScreen.js
@@ -1,48 +1,18 @@
 import React, { useState, useEffect } from 'react';
 import { View, Text, StyleSheet, Image, TouchableOpacity, Linking, Alert } from 'react-native';
-import CustomInput from '../components/CustomInput';
-import CustomButton from '../components/CustomButton';
 import axios from 'axios';
 import { WebView } from 'react-native-webview';
 import styles from '../components/LoginScreenStyles';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { getFCMToken, saveTokenToServer } from '../utils/FCMUtils';
 import { API_URL } from '@env';
-import { useNavigation, useFocusEffect } from '@react-navigation/native';
+import { useNavigation } from '@react-navigation/native';
 
-const LoginScreen = ({ }) => {
+const LoginScreen = () => {
     const [responseMessage, setResponseMessage] = useState('');
     const [loginUrl, setLoginUrl] = useState(null);
+    const [hasNavigated, setHasNavigated] = useState(false);
     const navigation = useNavigation();
-
-    const handleLogin = async () => {
-        try {
-            const response = await axios.post(`${API_URL}/auth/login`, {
-                userId: userId,
-                userPw: userPw,
-            });
-            const accessToken = response.headers.get("accessToken");
-            const userSeq = response.headers.get("userSeq");
-            await AsyncStorage.setItem('accessToken', accessToken);
-            await AsyncStorage.setItem('userSeq', userSeq);
-            // 로그인 성공 후 FCM 토큰 저장
-            const token = await getFCMToken();
-            if (token) {
-                // 서버에 토큰 저장 (userId는 로그인 후에 얻은 값으로 사용)
-                console.log(token);
-                saveTokenToServer(token, userId);  // 여기서 userId를 서버로 보냄
-            }
-
-            setResponseMessage(`Login successful: ${response.data.message}`);
-            navigation.navigate('Board');
-        } catch (error) {
-            if (error.response) {
-                setResponseMessage(`Login failed: ${error.response.data.error}`);
-            } else {
-                setResponseMessage('Error connecting to the server');
-            }
-        }
-    };
 
     const handleKakaoLogin = async () => {
         try {
@@ -58,37 +28,32 @@ const LoginScreen = ({ }) => {
         }
     };
 
-    useEffect(() => {
-        if (loginUrl) {
-            // When login URL changes, handle redirection logic
-        }
-    }, [loginUrl]);
-
     if (loginUrl) {
         return (
             <WebView
-                source={{ uri: loginUrl, headers: {} }}
+                source={{ uri: loginUrl }}
                 style={{ flex: 1 }}
                 onNavigationStateChange={async (navState) => {
+                    if (hasNavigated) return; // 중복 상태 변경 방지
 
                     console.log("URL:", navState.url);
 
                     const accessTokenMatch = navState.url.match(/accessToken=([^&]+)/);
                     const userSeqMatch = navState.url.match(/userSeq=([^&]+)/);
-                    console.log("UserId:", userSeqMatch)
-                    if (accessTokenMatch && userSeqMatch && navState.canGoBack === true) {
-                        const accessToken = accessTokenMatch[1];
 
+                    if (accessTokenMatch && userSeqMatch) {
+                        const accessToken = accessTokenMatch[1];
                         const userSeq = userSeqMatch[1];
 
                         console.log("Access Token extracted:", accessToken);
-                        console.log("userSeq extracted:", userSeq);
+                        console.log("UserSeq extracted:", userSeq);
 
-                        await AsyncStorage.setItem('accessToken', "Bearer "+accessToken);
+                        await AsyncStorage.setItem('accessToken', `Bearer ${accessToken}`);
                         await AsyncStorage.setItem('userSeq', userSeq);
 
                         console.log("Access Token stored:", accessToken);
 
+                        setHasNavigated(true); // 상태 변경 플래그 설정
                         setLoginUrl(null);
                         console.log("Navigating to home");
                         navigation.navigate('Board');
@@ -110,7 +75,7 @@ const LoginScreen = ({ }) => {
 
             {/* 시작하기 버튼 */}
             <View style={styles.startButtonContainer}>
-                <TouchableOpacity onPress={navigation.navigate('StartLogin')} style={[styles.startButton, styles.customButton]}>
+                <TouchableOpacity onPress={() => navigation.navigate('StartLogin')} style={[styles.startButton, styles.customButton]}>
                     <Text style={styles.startButtonText}>시작하기</Text>
                 </TouchableOpacity>
             </View>
@@ -129,9 +94,9 @@ const LoginScreen = ({ }) => {
                     <Text style={styles.signUpLinkTextUnderLine}>회원가입</Text>
                 </Text>
             </TouchableOpacity>
-
         </View>
     );
 };
+
 
 export default LoginScreen;
