+++ src/screens/StartLogin.js
@@ -4,28 +4,39 @@ import { View, Text, TextInput, TouchableOpacity, StyleSheet, Alert } from 'reac
 import CustomButton from '../components/CustomButton';
 import axios from 'axios';
 import { API_URL } from '@env';
+import AsyncStorage from '@react-native-async-storage/async-storage';
+import { getFCMToken, saveTokenToServer } from '../utils/FCMUtils';
 
 const StartLogin = ({ navigation }) => {
     const [userId, setUserId] = useState('');
     const [userPw, setUserPw] = useState('');
-
+    const [responseMessage, setResponseMessage] = useState('');
+    
     const handleLogin = async () => {
-        if (!userId || !userPw) {
-            Alert.alert('Error', 'Please enter both ID and Password.');
-            return;
-        }
         try {
             const response = await axios.post(`${API_URL}/auth/login`, {
                 userId: userId,
                 userPw: userPw,
             });
-            Alert.alert('Success', `Welcome, ${userId}!`);
-            navigation.navigate('Home'); // 로그인 성공 후 홈 화면으로 이동
+            const accessToken = response.headers.get("accessToken");
+            const userSeq = response.headers.get("userSeq");
+            await AsyncStorage.setItem('accessToken', accessToken);
+            await AsyncStorage.setItem('userSeq', userSeq);
+            // 로그인 성공 후 FCM 토큰 저장 
+            const token = await getFCMToken();
+            if (token) {
+                // 서버에 토큰 저장 (userId는 로그인 후에 얻은 값으로 사용)
+                console.log(token);
+                saveTokenToServer(token, userId);  // 여기서 userId를 서버로 보냄
+            }
+
+            setResponseMessage(`Login successful: ${response.data.message}`);
+            navigation.navigate('Board');
         } catch (error) {
             if (error.response) {
-                Alert.alert('Login failed', error.response.data.error);
+                setResponseMessage(`Login failed: ${error.response.data.error}`);
             } else {
-                Alert.alert('Error', 'Unable to connect to the server');
+                setResponseMessage('Error connecting to the server');
             }
         }
     };
